Abandon multiboot in favour of Cursed Bootloader
Cursed Bootloader is written in 16-bit real mode x86, is extremely
bugprone and lacking in features, but manages to boot this thing on
actual hardware.
It has the distinct advantage of fitting into a boot sector, which means
I can target my ancient hardware. Currently targeting my 486DX2-66 with
32MB of RAM :]

The boot sector code uses BIOS interrupts to load the raw elf image into
RAM, and then we set up protected mode and jump to and offset in there
to get to _start(). This was inspired by the way I saw it being done in
SerenityOS back in the day. Needless to say, this is very much not
recommended.