meta: Add bochs target for kernel
Bochs is a bit cumbersome to run, and I didn't have any notes from when
I last used it, so the makefile now documents a bit how to use it.
The new 'bochs' target does require a suitably configured copy of bochs
be placed under ~/git/Bochs manually, at least for now.